RoomList 컴포넌트 리팩토링 및 Item, List 컴포넌트 분리
